// src/types/footer.ts
export interface Footer {
  descricao: string
  email: string
  Telefone: string
  link_youtube: {
    title: string
    uri: string
  }
  link_instagram: {
    title: string
    uri: string
  }
  link_linkedin: {
    title: string
    uri: string
  }
  link_facebook: {
    title: string
    uri: string
  }
  logo: string
}